+++ packages/creamy_field/CHANGELOG.md
@@ -1,2 +1,11 @@
 # Changelog
-## [0.1.0] - 4 July 2020
+
+## 0.2.0
+
+- Modified Text controller
+- Modified new syntax highligher as an implementation of `SyntaxHighlighter` and renamed it with prefix Creamy
+
+## 0.1.0
+
+- CreamyField to be used as a Rich Editing Text Field
+- Syntax highlighting support